+++ packages/bloc/lib/bloc.dart
@@ -5,5 +5,4 @@ library bloc;
 
 export './src/bloc.dart';
 export './src/bloc_observer.dart';
-export './src/bloc_overrides.dart';
 export './src/transition.dart';
