+++ packages/bloc/lib/src/bloc.dart
@@ -3,6 +3,97 @@ import 'dart:async';
 import 'package:bloc/bloc.dart';
 import 'package:meta/meta.dart';
 
+const _asyncRunZoned = runZoned;
+
+/// {@template bloc_overrides}
+/// This class facilitates overriding [BlocObserver] and [EventTransformer].
+/// It should be extended by another class in client code with overrides
+/// that construct a custom implementation. The implementation in this class
+/// defaults to the base [blocObserver] and [eventTransformer] implementation.
+/// For example:
+///
+/// ```dart
+/// class MyBlocObserver extends BlocObserver {
+///   ...
+///   // A custom BlocObserver implementation.
+///   ...
+/// }
+///
+/// void main() {
+///   BlocOverrides.runZoned(() {
+///     ...
+///     // Bloc instances will use MyBlocObserver instead of the default BlocObserver.
+///     ...
+///   }, blocObserver: MyBlocObserver());
+/// }
+/// ```
+/// {@endtemplate}
+abstract class BlocOverrides {
+  /// {@macro bloc_overrides}
+  const BlocOverrides();
+
+  static final _token = Object();
+
+  /// Returns the current [BlocOverrides] instance.
+  ///
+  /// This will return `null` if the current [Zone] does not contain
+  /// any [BlocOverrides].
+  ///
+  /// See also:
+  /// * [BlocOverrides.runZoned] to provide [BlocOverrides]
+  /// in a fresh [Zone].
+  ///
+  static BlocOverrides? get current {
+    return Zone.current[_token] as BlocOverrides?;
+  }
+
+  /// Runs [body] in a fresh [Zone] using the provided overrides.
+  static R runZoned<R>(
+    R Function() body, {
+    BlocObserver? blocObserver,
+    EventTransformer? eventTransformer,
+  }) {
+    final overrides = _BlocOverridesScope(blocObserver, eventTransformer);
+    return _asyncRunZoned(body, zoneValues: {_token: overrides});
+  }
+
+  /// The default [BlocObserver] instance.
+  BlocObserver get blocObserver => Bloc._defaultBlocObserver;
+
+  /// The default [EventTransformer] used for all event handlers.
+  /// By default all events are processed concurrently.
+  ///
+  /// If a custom transformer is specified for a particular event handler,
+  /// it will take precendence over the global transformer.
+  ///
+  /// See also:
+  ///
+  /// * [package:bloc_concurrency](https://pub.dev/packages/bloc_concurrency) for an
+  /// opinionated set of event transformers.
+  ///
+  EventTransformer get eventTransformer => Bloc._defaultEventTransformer;
+}
+
+class _BlocOverridesScope extends BlocOverrides {
+  _BlocOverridesScope(this._blocObserver, this._eventTransformer);
+
+  final BlocOverrides? _previous = BlocOverrides.current;
+  final BlocObserver? _blocObserver;
+  final EventTransformer? _eventTransformer;
+
+  @override
+  BlocObserver get blocObserver {
+    return _blocObserver ?? _previous?.blocObserver ?? super.blocObserver;
+  }
+
+  @override
+  EventTransformer get eventTransformer {
+    return _eventTransformer ??
+        _previous?.eventTransformer ??
+        super.eventTransformer;
+  }
+}
+
 /// {@template emitter}
 /// An [Emitter] is a class which is capable of emitting new states.
 ///
@@ -224,6 +315,15 @@ abstract class Bloc<Event, State> extends BlocBase<State> {
   final _subscriptions = <StreamSubscription<dynamic>>[];
   final _handlers = <_Handler>[];
   final _emitters = <_Emitter>[];
+  final _eventTransformer =
+      BlocOverrides.current?.eventTransformer ?? _defaultEventTransformer;
+
+  static final _defaultBlocObserver = BlocObserver();
+  static final _defaultEventTransformer = (Stream events, EventMapper mapper) {
+    return events
+        .map(mapper)
+        .transform<dynamic>(const _FlatMapStreamTransformer<dynamic>());
+  };
 
   /// Notifies the [Bloc] of a new [event] which triggers
   /// all corresponding [EventHandler] instances.
@@ -276,7 +376,7 @@ abstract class Bloc<Event, State> extends BlocBase<State> {
   @mustCallSuper
   void onEvent(Event event) {
     // ignore: invalid_use_of_protected_member
-    BlocOverrides.current.blocObserver.onEvent(this, event);
+    _blocObserver.onEvent(this, event);
   }
 
   /// {@template emit}
@@ -344,7 +444,7 @@ abstract class Bloc<Event, State> extends BlocBase<State> {
       return true;
     }());
 
-    final _transformer = transformer ?? BlocOverrides.current.eventTransformer;
+    final _transformer = transformer ?? _eventTransformer;
     final subscription = _transformer(
       _eventController.stream.where((event) => event is E).cast<E>(),
       (dynamic event) {
@@ -416,7 +516,7 @@ abstract class Bloc<Event, State> extends BlocBase<State> {
   @mustCallSuper
   void onTransition(Transition<Event, State> transition) {
     // ignore: invalid_use_of_protected_member
-    BlocOverrides.current.blocObserver.onTransition(this, transition);
+    _blocObserver.onTransition(this, transition);
   }
 
   /// Closes the `event` and `state` `Streams`.
@@ -467,8 +567,11 @@ abstract class BlocBase<State> {
   /// {@macro bloc_stream}
   BlocBase(this._state) {
     // ignore: invalid_use_of_protected_member
-    BlocOverrides.current.blocObserver.onCreate(this);
+    _blocObserver.onCreate(this);
   }
+
+  final _blocObserver =
+      BlocOverrides.current?.blocObserver ?? Bloc._defaultBlocObserver;
 
   StreamController<State>? __stateController;
   StreamController<State> get _stateController {
@@ -540,7 +643,7 @@ abstract class BlocBase<State> {
   @mustCallSuper
   void onChange(Change<State> change) {
     // ignore: invalid_use_of_protected_member
-    BlocOverrides.current.blocObserver.onChange(this, change);
+    _blocObserver.onChange(this, change);
   }
 
   /// Reports an [error] which triggers [onError] with an optional [StackTrace].
@@ -566,7 +669,7 @@ abstract class BlocBase<State> {
   @mustCallSuper
   void onError(Object error, StackTrace stackTrace) {
     // ignore: invalid_use_of_protected_member
-    BlocOverrides.current.blocObserver.onError(this, error, stackTrace);
+    _blocObserver.onError(this, error, stackTrace);
   }
 
   /// Closes the instance.
@@ -575,7 +678,52 @@ abstract class BlocBase<State> {
   @mustCallSuper
   Future<void> close() async {
     // ignore: invalid_use_of_protected_member
-    BlocOverrides.current.blocObserver.onClose(this);
+    _blocObserver.onClose(this);
     await _stateController.close();
+  }
+}
+
+class _FlatMapStreamTransformer<T> extends StreamTransformerBase<Stream<T>, T> {
+  const _FlatMapStreamTransformer();
+
+  @override
+  Stream<T> bind(Stream<Stream<T>> stream) {
+    final controller = StreamController<T>.broadcast(sync: true);
+
+    controller.onListen = () {
+      final subscriptions = <StreamSubscription<dynamic>>[];
+
+      final outerSubscription = stream.listen(
+        (inner) {
+          final subscription = inner.listen(
+            controller.add,
+            onError: controller.addError,
+          );
+
+          subscription.onDone(() {
+            subscriptions.remove(subscription);
+            if (subscriptions.isEmpty) controller.close();
+          });
+
+          subscriptions.add(subscription);
+        },
+        onError: controller.addError,
+      );
+
+      outerSubscription.onDone(() {
+        subscriptions.remove(outerSubscription);
+        if (subscriptions.isEmpty) controller.close();
+      });
+
+      subscriptions.add(outerSubscription);
+
+      controller.onCancel = () {
+        if (subscriptions.isEmpty) return null;
+        final cancels = [for (final s in subscriptions) s.cancel()];
+        return Future.wait(cancels).then((_) {});
+      };
+    };
+
+    return controller.stream;
   }
 }
